pyproject.toml
[project]
name = "coding-standards"
version = "0.0.0"
requires-python = ">=3.11"
dependencies = [
    "networkx>=3.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-randomly",
]

[tool.pytest.ini_options]
pythonpath = ["."]

[tool.deptry]
# manifest_schema: local module in scripts/, not an installable package
known_first_party = ["manifest_schema", "blast_radius"]
# pyyaml installs as 'yaml', pytest-randomly activates via entry point
package_module_name_map = {pyyaml = "yaml"}
# typer: only used inside Docker (entrypoint.py), not a local dev dependency
per_rule_ignores = {DEP002 = ["pytest-randomly"], DEP003 = ["typer"]}

[tool.importlinter]
root_packages = ["scripts"]
# Scripts are standalone — no layered architecture to enforce yet.
# Minimal contract validates import-linter runs without "no configuration" error.
